Move test classes into correct package
